+++ geolocator/lib/geolocator.dart
@@ -35,6 +35,14 @@ class Geolocator {
   static Future<bool> isLocationServiceEnabled() =>
       GeolocatorPlatform.instance.isLocationServiceEnabled();
 
+  /// Returns a [Future] containing a [bool] value indicating whether the
+  /// Google Location Accuracy for the FusedLocationProvider is enabled on
+  /// the device.
+  /// Returns hardcoded `false` on iOS.
+  /// Returns hardcoded `false` when Google Play Services are not installed.
+  static Future<bool> isGoogleLocationAccuracyEnabled() =>
+      GeolocatorPlatform.instance.isGoogleLocationAccuracyEnabled();
+
   /// Returns the last known position stored on the users device.
   ///
   /// On Android you can force the plugin to use the old Android