+++ geolocator_web/lib/src/geolocation_manager.dart
@@ -1,5 +1,6 @@
 import 'package:geolocator_platform_interface/geolocator_platform_interface.dart';
 
+// ignore_for_file: public_member_api_docs
 abstract class GeolocationManager {
   Future<Position> getCurrentPosition({
     bool? enableHighAccuracy,